refactor(brutus): extract worker pool engine to workers.go
