+++ pkg/brutus/workers.go
@@ -1,0 +1,350 @@
+// Copyright 2026 Praetorian Security, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package brutus
+
+import (
+	"context"
+	"math/rand"
+	"sync"
+	"sync/atomic"
+	"time"
+
+	"golang.org/x/sync/errgroup"
+	"golang.org/x/time/rate"
+)
+
+// credential represents a single authentication attempt.
+type credential struct {
+	username     string
+	password     string
+	key          []byte // SSH private key (optional, for key-based auth)
+	llmSuggested bool   // True if this credential was suggested by LLM
+}
+
+// generateCredentials creates all possible username/password combinations.
+func generateCredentials(usernames, passwords []string) []credential {
+	creds := make([]credential, 0, len(usernames)*len(passwords))
+
+	for _, username := range usernames {
+		for _, password := range passwords {
+			creds = append(creds, credential{
+				username: username,
+				password: password,
+			})
+		}
+	}
+
+	return creds
+}
+
+// generateKeyCredentials creates all possible username/key combinations.
+func generateKeyCredentials(usernames []string, keys [][]byte) []credential {
+	creds := make([]credential, 0, len(usernames)*len(keys))
+
+	for _, username := range usernames {
+		for _, key := range keys {
+			creds = append(creds, credential{
+				username: username,
+				key:      key,
+			})
+		}
+	}
+
+	return creds
+}
+
+// reorderForSpray reorders credentials to try each password across all users
+// before moving to the next password (password spraying mode).
+func reorderForSpray(creds []credential) []credential {
+	if len(creds) == 0 {
+		return creds
+	}
+
+	// Group credentials by password
+	byPassword := make(map[string][]credential)
+	passwordOrder := []string{}
+
+	for _, c := range creds {
+		if _, seen := byPassword[c.password]; !seen {
+			passwordOrder = append(passwordOrder, c.password)
+		}
+		byPassword[c.password] = append(byPassword[c.password], c)
+	}
+
+	// Rebuild credentials list: all users for password1, then all users for password2, etc.
+	result := make([]credential, 0, len(creds))
+	for _, pass := range passwordOrder {
+		result = append(result, byPassword[pass]...)
+	}
+
+	return result
+}
+
+// runWorkers executes credential testing using a bounded worker pool.
+func runWorkers(ctx context.Context, cfg *Config, plug Plugin) ([]Result, error) {
+	// Add TLSMode to context at the start
+	ctx = ContextWithTLSMode(ctx, cfg.TLSMode)
+
+	// Check if LLM analysis is enabled AND protocol supports it
+	// LLM banner analysis only makes sense for HTTP Basic Auth where we can
+	// detect the application from the response headers/body
+	if cfg.LLMConfig != nil && cfg.LLMConfig.Enabled && isHTTPProtocol(cfg.Protocol) {
+		// Use LLM-enhanced flow: capture banner, analyze, test suggestions
+		return runWorkersWithLLM(ctx, cfg, plug)
+	}
+
+	// Default flow: test credentials without LLM analysis
+	return runWorkersDefault(ctx, cfg, plug)
+}
+
+// isHTTPProtocol returns true if the protocol uses HTTP Basic Auth
+// and can benefit from LLM-based application detection.
+func isHTTPProtocol(protocol string) bool {
+	switch protocol {
+	case "http", "https", "couchdb", "elasticsearch", "influxdb":
+		return true
+	default:
+		return false
+	}
+}
+
+// executeWorkerPool is the shared worker pool implementation used by both
+// runWorkersDefault and runWorkersWithLLM. It handles concurrency control,
+// rate limiting, jitter, max attempts, result collection, and early stopping.
+func executeWorkerPool(ctx context.Context, cfg *Config, plug Plugin, credentials []credential, llmSuggestions []string) ([]Result, error) {
+	// Create cancellable context for early stop
+	ctx, cancel := context.WithCancel(ctx)
+	defer cancel()
+
+	// Setup errgroup with bounded concurrency
+	g, ctx := errgroup.WithContext(ctx)
+	g.SetLimit(cfg.Threads)
+
+	// Create rate limiter if configured
+	var limiter *rate.Limiter
+	if cfg.RateLimit > 0 {
+		limiter = rate.NewLimiter(rate.Limit(cfg.RateLimit), 1)
+	}
+
+	// Result collection with mutex protection
+	var (
+		results       []Result
+		attemptCounts = make(map[string]int)
+		attemptMu     sync.Mutex
+		mu            sync.Mutex
+		found         atomic.Bool
+	)
+
+	// Launch workers
+	for _, cred := range credentials {
+		// Check early stop before launching new worker
+		if cfg.StopOnSuccess && found.Load() {
+			break
+		}
+
+		// Capture loop variable for closure
+		cred := cred
+
+		g.Go(func() error {
+			// Check context cancellation
+			select {
+			case <-ctx.Done():
+				return nil
+			default:
+			}
+
+			// Apply rate limiting if configured
+			if limiter != nil {
+				if err := limiter.Wait(ctx); err != nil {
+					return nil // Context canceled
+				}
+				// Apply jitter if configured
+				if cfg.Jitter > 0 {
+					jitterDuration := time.Duration(rand.Int63n(int64(cfg.Jitter)))
+					select {
+					case <-time.After(jitterDuration):
+						// Jitter sleep completed
+					case <-ctx.Done():
+						return nil // Context canceled during jitter
+					}
+				}
+			}
+
+			// Check max attempts per user
+			if cfg.MaxAttempts > 0 {
+				attemptMu.Lock()
+				if attemptCounts[cred.username] >= cfg.MaxAttempts {
+					attemptMu.Unlock()
+					return nil
+				}
+				attemptCounts[cred.username]++
+				attemptMu.Unlock()
+			}
+
+			// Test credential (key-based or password-based)
+			var result *Result
+			if cred.key != nil {
+				// Key-based authentication
+				// Check if plugin supports key auth
+				if kp, ok := plug.(KeyPlugin); ok {
+					result = kp.TestKey(ctx, cfg.Target, cred.username, cred.key, cfg.Timeout)
+				} else {
+					// Plugin doesn't support key auth, skip
+					return nil
+				}
+			} else {
+				// Password-based authentication
+				result = plug.Test(ctx, cfg.Target, cred.username, cred.password, cfg.Timeout)
+			}
+
+			// Populate LLM tracking fields if suggestions were provided
+			if len(llmSuggestions) > 0 {
+				result.LLMSuggested = cred.llmSuggested
+				result.LLMSuggestedCreds = llmSuggestions
+			}
+
+			// Collect result
+			mu.Lock()
+			results = append(results, *result)
+			mu.Unlock()
+
+			// Handle success with early stop
+			if result.Success && cfg.StopOnSuccess {
+				found.Store(true)
+				cancel() // Signal all workers to stop
+			}
+
+			return nil
+		})
+	}
+
+	// Wait for all workers to complete
+	if err := g.Wait(); err != nil && err != context.Canceled {
+		return results, err
+	}
+
+	return results, nil
+}
+
+// runWorkersDefault executes credential testing using a bounded worker pool.
+// Uses errgroup for concurrency control and context cancellation for early stopping.
+func runWorkersDefault(ctx context.Context, cfg *Config, plug Plugin) ([]Result, error) {
+	// Generate all credential combinations
+	var credentials []credential
+
+	// Add pre-paired credentials (no Cartesian product)
+	for _, c := range cfg.Credentials {
+		credentials = append(credentials, credential{
+			username: c.Username,
+			password: c.Password,
+			key:      c.Key,
+		})
+	}
+
+	// Add password-based credentials (Cartesian product)
+	if len(cfg.Passwords) > 0 {
+		credentials = append(credentials, generateCredentials(cfg.Usernames, cfg.Passwords)...)
+	}
+
+	// Add key-based credentials (Cartesian product, if supported by plugin)
+	if len(cfg.Keys) > 0 {
+		credentials = append(credentials, generateKeyCredentials(cfg.Usernames, cfg.Keys)...)
+	}
+
+	// Reorder for spray mode if enabled
+	if cfg.SprayMode {
+		credentials = reorderForSpray(credentials)
+	}
+
+	// Execute worker pool with no LLM suggestions
+	return executeWorkerPool(ctx, cfg, plug, credentials, nil)
+}
+
+// runWorkersWithLLM executes credential testing with optional LLM-based banner analysis.
+// Phase 1: Capture banner with dummy credential
+// Phase 2: Check if standard banner (skip LLM if standard)
+// Phase 3: Analyze non-standard banner with LLM
+// Phase 4: Test LLM suggestions first (priority)
+// Phase 5: Test default credentials
+// Phase 6: Run workers with combined credential list
+func runWorkersWithLLM(ctx context.Context, cfg *Config, plug Plugin) ([]Result, error) {
+	// Phase 1: Capture banner with dummy credential
+	banner := captureBanner(ctx, cfg, plug)
+
+	// Phase 2: Check if standard banner
+	if IsStandardBanner(cfg.Protocol, banner.Banner) {
+		// Standard banner - use default credentials only
+		return runWorkersDefault(ctx, cfg, plug)
+	}
+
+	// Phase 3: Non-standard banner - get LLM suggestions
+	analyzer := createAnalyzer(cfg.LLMConfig)
+	if analyzer == nil {
+		// Analyzer creation failed - fallback to defaults
+		return runWorkersDefault(ctx, cfg, plug)
+	}
+
+	suggestions, err := analyzer.Analyze(ctx, banner)
+	if err != nil {
+		// LLM analysis failed - fallback to defaults
+		return runWorkersDefault(ctx, cfg, plug)
+	}
+
+	// Phase 4: Build LLM credential list (test these first)
+	llmCreds := []credential{}
+	for _, username := range cfg.Usernames {
+		for _, password := range suggestions {
+			llmCreds = append(llmCreds, credential{
+				username:     username,
+				password:     password,
+				llmSuggested: true,
+			})
+		}
+	}
+
+	// Phase 5: Build default credential list
+	defaultCreds := generateCredentials(cfg.Usernames, cfg.Passwords)
+
+	// Phase 6: Combine LLM suggestions first, then defaults
+	allCreds := make([]credential, 0, len(llmCreds)+len(defaultCreds))
+	allCreds = append(allCreds, llmCreds...)
+	allCreds = append(allCreds, defaultCreds...)
+
+	// Run workers with combined credentials
+	return executeWorkerPool(ctx, cfg, plug, allCreds, suggestions)
+}
+
+// captureBanner makes an initial connection to capture the service banner.
+// Uses a dummy credential to trigger the connection and extract banner information.
+func captureBanner(ctx context.Context, cfg *Config, plug Plugin) BannerInfo {
+	// Use first username with empty password for banner capture
+	// If no usernames provided (only pre-paired Credentials), extract from first Credential
+	var username string
+	if len(cfg.Usernames) > 0 {
+		username = cfg.Usernames[0]
+	} else if len(cfg.Credentials) > 0 {
+		username = cfg.Credentials[0].Username
+	}
+	// Empty username is acceptable for banner capture (some protocols don't need it)
+
+	// Test with dummy credential just to capture banner
+	result := plug.Test(ctx, cfg.Target, username, "", cfg.Timeout)
+
+	return BannerInfo{
+		Protocol: cfg.Protocol,
+		Target:   cfg.Target,
+		Banner:   result.Banner,
+	}
+}
